agent_tutorial/introduction/basic/requirements.txt
llama-index>=0.10.0
python-dotenv
sentence-transformers
torch
llama-index-embeddings-huggingface
llama-index-llms-deepseek